test(mobile): extract shared makeRaw* factories to __tests__/factories.ts
Consolidates 28 duplicated makeRaw<Entity> factories from 23 mobile hook
tests into a single shared module. Each factory accepts a
Partial<EntityRaw> overrides argument so test-specific tweaks no longer
require re-declaring the whole fixture. TEST_MASTER_KEY / TEST_SYSTEM_ID
are re-exported from the canonical helpers/test-crypto module to preserve
the single source of truth for round-tripping. Eight factories renamed
for disambiguation in the shared namespace.

Completes mobile-8ovj.
