+++ apps/mobile/src/hooks/__tests__/use-groups.test.tsx
@@ -1,19 +1,15 @@
 // @vitest-environment happy-dom
 import { configureSodium, initSodium } from "@pluralscape/crypto";
 import { WasmSodiumAdapter } from "@pluralscape/crypto/wasm";
-import { encryptGroupInput } from "@pluralscape/data/transforms/group";
 import { brandId } from "@pluralscape/types";
 import { act, waitFor } from "@testing-library/react";
 import { beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
 
-import {
-  renderHookWithProviders,
-  TEST_MASTER_KEY,
-  TEST_SYSTEM_ID,
-} from "./helpers/render-hook-with-providers.js";
+import { makeRawGroup } from "../../__tests__/factories.js";
 
-import type { GroupRaw } from "@pluralscape/data/transforms/group";
-import type { GroupId, UnixMillis } from "@pluralscape/types";
+import { renderHookWithProviders, TEST_SYSTEM_ID } from "./helpers/render-hook-with-providers.js";
+
+import type { GroupId } from "@pluralscape/types";
 
 beforeAll(async () => {
   configureSodium(new WasmSodiumAdapter());
@@ -132,34 +128,6 @@ const {
   useRemoveGroupMembers,
   useReorderGroups,
 } = await import("../use-groups.js");
-
-// ── Fixtures ─────────────────────────────────────────────────────────
-const NOW = 1_700_000_000_000 as UnixMillis;
-
-function makeRawGroup(id: string): GroupRaw {
-  const encrypted = encryptGroupInput(
-    {
-      name: `Group ${id}`,
-      description: "A test group",
-      imageSource: null,
-      color: null,
-      emoji: null,
-    },
-    TEST_MASTER_KEY,
-  );
-  return {
-    id: brandId<GroupId>(id),
-    systemId: TEST_SYSTEM_ID,
-    parentGroupId: null,
-    sortOrder: 0,
-    version: 1,
-    createdAt: NOW,
-    updatedAt: NOW,
-    archived: false,
-    archivedAt: null,
-    ...encrypted,
-  };
-}
 
 beforeEach(() => {
   fixtures.clear();
